requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
requests
python-dotenv
scikit-learn
faiss-cpu
sentence-transformers
pandas
openpyxl
numpy
streamlit
